refactor code for exec coming later
